test update products ok
